Utworzenie tabel `cars` oraz `orders`

+++ app/src/main/java/com/rental/car/carrentalbeaverandroid/dbconnection/DatabaseConfig.java
@@ -10,7 +10,7 @@ import com.rental.car.carrentalbeaverandroid.models.User;
 public class DatabaseConfig extends SQLiteOpenHelper {
 
     public DatabaseConfig(Context context){
-        super(context, "moja-baza.db",null,2);
+        super(context, "moja-baza.db",null,3);
 
     }
 
@@ -30,8 +30,31 @@ public class DatabaseConfig extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE users;");
-        onCreate(db);
+        if(oldVersion<2) {
+            db.execSQL("DROP TABLE users;");
+            onCreate(db);
+        }
+        if(oldVersion<3){
+            db.execSQL("CREATE UNIQUE INDEX user_email_unique_index ON users(user_email)");
+
+            db.execSQL("CREATE TABLE cars (" +
+                    "car_id integer primary key autoincrement," +
+                    "car_name text not null," +
+                    "car_price DECIMAL(6,2) not null" +
+                    ")");
+
+            db.execSQL("CREATE TABLE orders(" +
+                    " order_id integer primary key autoincrement, " +
+                    " order_user integer not null, " +
+                    " order_car integer not null, " +
+                    " order_start_date text not null, " +
+                    " order_end_date text not null, " +
+                    " FOREIGN KEY (order_user) REFERENCES users (user_id) " +
+                    " ON DELETE CASCADE ON UPDATE NO ACTION, " +
+                    " FOREIGN KEY (order_car) REFERENCES cars (car_id) " +
+                    " ON DELETE CASCADE ON UPDATE NO ACTION" +
+                    ")");
+        }
     }
 
     private void enterUser(SQLiteDatabase db, String email, String password){
